+++ CHANGELOG.md
@@ -4,6 +4,10 @@ We aim to adhere to [semantic versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Version 0.2.1 - Unreleased
 
+### Changed
+
+* YAML paths in grid values no longer auto-expand unless explicitly behind an `__include__` key. See docs for details.
+
 
 ## [Version 0.0.1] -